LPP-60517 Add CDN to ES CSS loaders
